Dynamic Body Studio Mobile Application
Here resides all the java and XML code that was used to create the final working version of the mobile application. The app was designed and developed for a local company as a third-year project for university. The company is registered under the name of "Dynamic Body Studio", which provides fitness classes to the city of East London.