Hotel Management System
In the second year of university, we were tasked to develop a system which included a registration functionality. My group designed and developed a hotel management system, with its own stock control system, room service, registration and much more. Here are the java files that contain the code that was used to build the system.